+++ src/background/background-request.ts
@@ -9,15 +9,15 @@ const SearchRequestSchema = s.type({
 export type SearchRequest = s.Infer<typeof SearchRequestSchema>;
 
 export const BackgroundRequestSchema = discriminator('type', {
+  //
+  // Requests for the background page
+  //
   disabled: s.type({}),
   'enable?': s.type({}),
   enabled: s.type({
     src: s.string(),
   }),
-  //
-  // Requests for the background page
-  //
-
+  forceGdocsHtmlMode: s.type({}),
   options: s.type({}),
   // TODO: Remove 'search' once we have shipped the two-step search approach.
   //