+++ src/options/options.ts
@@ -181,6 +181,12 @@ function completeForm() {
   });
 
   document
+    .getElementById('forceGdocsHtmlMode')!
+    .addEventListener('click', (evt) => {
+      config.forceGdocsHtmlMode = (evt.target as HTMLInputElement).checked;
+    });
+
+  document
     .getElementById('showKanjiComponents')!
     .addEventListener('click', (evt) => {
       config.showKanjiComponents = (evt.target as HTMLInputElement).checked;
@@ -870,6 +876,7 @@ function fillVals() {
   optform.tabDisplay.value = config.tabDisplay;
   optform.toolbarIcon.value = config.toolbarIcon;
   optform.showPuck.value = config.showPuck;
+  optform.forceGdocsHtmlMode.checked = config.forceGdocsHtmlMode;
 
   renderCurrencyList(config.fxCurrency, config.fxCurrencies);